+++ credmark/cmf/types/token.py
@@ -1,48 +1,55 @@
 from typing import List, Union
-from credmark.dto import PrivateAttr, IterableListGenericDTO, DTOField
-import credmark.cmf.model
-from credmark.cmf.model.errors import ModelDataError, ModelRunError
-from credmark.cmf.types.currency import Currency
-
-from .abi import ABI
-from .contract import Contract
-from .address import NATIVE_TOKEN_ADDRESS, Address
-from .data.fungible_token_data import FUNGIBLE_TOKEN_DATA, ERC20_GENERIC_ABI
 from web3.exceptions import (
     BadFunctionCallOutput,
     ABIFunctionNotFound
 )
+
+from credmark.dto import PrivateAttr, IterableListGenericDTO, DTOField, DTO
+import credmark.cmf.model
+from credmark.cmf.model.errors import ModelDataError, ModelRunError
+
+from .abi import ABI
+from .address import Address
+from .account import Account
+from .contract import Contract
+from .data.fungible_token_data import (
+    NATIVE_TOKEN,
+    FUNGIBLE_TOKEN_DATA_BY_SYMBOL,
+    FUNGIBLE_TOKEN_DATA_BY_ADDRESS
+)
+from .data.erc_standard_data import ERC20_BASE_ABI
+from .data.fiat_currency_data import FIAT_CURRENCY_DATA_BY_SYMBOL, FIAT_CURRENCY_DATA_BY_ADDRESS
 
 
 def get_token_from_configuration(
         chain_id: str,
         symbol: Union[str, None] = None,
         address: Union[Address, None] = None,
-        is_native_token: bool = False,
-        wraps: Union[str, None] = None) -> Union[dict, None]:
-    token_datas = [
-        t for t in FUNGIBLE_TOKEN_DATA.get(chain_id, [])
-        if (t.get('is_native_token', False) == is_native_token and
-            (t.get('address', None) == address or
-             t.get('symbol', None) == symbol) or
-            (t.get('wraps', None) == wraps and wraps is not None))
-    ]
+        is_native_token: bool = False) -> Union[dict, None]:
 
-    if len(token_datas) > 1:
-        # TODO: Until we have definitive token lookup, we'll
-        # consider it transient as a ModelRunError.
-        raise ModelRunError(
-            'Missing fungible token data in lookup for '
-            f'chain_id={chain_id} symbol={symbol} '
-            f'address={address} is_native_token={is_native_token}')
+    if is_native_token:
+        native_token_meta = NATIVE_TOKEN[chain_id]
+        if ((symbol is None or native_token_meta['symbol'] == symbol) and
+                (address is None or native_token_meta['address'] == Address(address))):
+            return native_token_meta
+        else:
+            return None
 
-    if len(token_datas) == 1:
-        return token_datas[0]
+    chain_tokens_by_symbol = FUNGIBLE_TOKEN_DATA_BY_SYMBOL.get(chain_id, {})
+    chain_tokens_by_address = FUNGIBLE_TOKEN_DATA_BY_ADDRESS.get(chain_id, {})
 
-    return None
+    if symbol is not None:
+        token_meta = chain_tokens_by_symbol.get(symbol, None)
+        if ((token_meta is not None) and
+                (address is None or token_meta['address'] == Address(address))):
+            return token_meta
+        else:
+            return None
+    else:
+        return chain_tokens_by_address.get(address, None)
 
 
-class Token(Contract, Currency):
+class Token(Contract):
     """
     Token represents a fungible Token that conforms to ERC20
     standards
@@ -64,6 +71,28 @@ class Token(Contract, Currency):
                          ] + Contract.Config.schema_extra['examples']
         }
 
+    def __new__(cls, **data):
+        if cls == NativeToken:
+            return super().__new__(cls)
+
+        context = credmark.cmf.model.ModelContext.current_context()
+        symbol = data.get('symbol', None)
+        address = data.get('address', None)
+
+        if symbol is None and address is None:
+            return super().__new__(cls)
+
+        token_data = get_token_from_configuration(
+            chain_id=str(context.chain_id),
+            symbol=symbol,
+            address=address,
+            is_native_token=True)
+
+        if token_data is not None:
+            return NativeToken()
+
+        return super().__new__(cls)
+
     def __init__(self, **data):
         if 'address' not in data and 'symbol' not in data:
             raise ModelDataError('One of address or symbol is required')
@@ -73,6 +102,7 @@ class Token(Contract, Currency):
 
             token_data = get_token_from_configuration(
                 chain_id=str(context.chain_id), symbol=data['symbol'])
+
             if token_data is None:
                 raise ModelDataError(f'Unsupported symbol: {data["symbol"]}')
 
@@ -82,6 +112,9 @@ class Token(Contract, Currency):
             data['meta']['symbol'] = token_data['symbol']
             data['meta']['name'] = token_data['name']
             data['meta']['decimals'] = token_data['decimals']
+
+            if token_data.get('set_loaded_true', False):  # Special case for BTC
+                self._loaded = True
 
         if data['address'] == Address.null():
             raise ModelDataError(f'NULL address ({Address.null()}) is not a valid Token Address')
@@ -93,7 +126,7 @@ class Token(Contract, Currency):
             return
 
         if self._meta.abi is None:
-            self._meta.abi = ABI(ERC20_GENERIC_ABI)
+            self._meta.abi = ABI(ERC20_BASE_ABI)
 
         super()._load()
 
@@ -155,8 +188,23 @@ class Token(Contract, Currency):
             self._meta.total_supply = self.try_erc20_property('totalSupply')
         return self._meta.total_supply
 
+    @property
+    def total_supply_scaled(self) -> float:
+        return self.scaled(self.total_supply)
+
     def scaled(self, value) -> float:
         return value / (10 ** self.decimals)
+
+    def balance_of(self, address: Address) -> int:
+        balance = self.functions.balanceOf(address).call()
+        return balance
+
+    def balance_of_scaled(self, address: Address) -> float:
+        return self.scaled(self.balance_of(address))
+
+    @property
+    def fiat(self) -> bool:
+        return False
 
 
 class TokenInfo(Token):
@@ -165,21 +213,46 @@ class TokenInfo(Token):
 
 class NativeToken(Token):
 
-    def __init__(self, **data) -> None:
+    def __init__(self, **kwargs) -> None:
         context = credmark.cmf.model.ModelContext.current_context()
-        data = {"address": NATIVE_TOKEN_ADDRESS}
-        super().__init__(**data)
+        token_data = NATIVE_TOKEN[str(context.chain_id)]
+        if token_data is None:
+            raise ModelRunError(f'No native token specified for chain id: {context.chain_id}')
+
+        symbol = kwargs.get('symbol', None)
+        address = kwargs.get('address', None)
+        if symbol is not None and token_data['symbol'] != symbol:
+            raise ModelRunError(
+                f'Wrong symbol {symbol} specified for {token_data["symbol"]} '
+                f'for chain id: {context.chain_id}')
+        if address is not None and token_data['address'] != Address(address):
+            raise ModelRunError(
+                f'Wrong address {address} specified for {token_data["address"]} '
+                f'for chain id: {context.chain_id}')
+        super().__init__(address=token_data['address'])
         if context.chain_id == 1:
             self._meta.abi = ABI([])
-            self._meta.symbol = "ETH"
-            self._meta.decimals = 18
-            self._meta.name = "Ethereum"
+            self._meta.symbol = token_data['symbol']
+            self._meta.name = token_data['name']
+            self._meta.decimals = token_data['decimals']
             self._meta.total_supply = 0
             self._loaded = True
 
-    def get_balance(self, address: Address) -> int:
+    def balance_of(self, address: Address) -> int:
         context = credmark.cmf.model.ModelContext.current_context()
-        return context.web3.eth.get_balance(address)
+        if context.chain_id == 1:
+            balance = context.web3.eth.get_balance(address)
+            return balance
+        else:
+            raise ModelRunError(f'Not supported for chain id: {context.chain_id}')
+
+    def balance_of_scaled(self, address: Address) -> float:
+        context = credmark.cmf.model.ModelContext.current_context()
+        if context.chain_id == 1:
+            balance = self.balance_of(address)
+            return float(context.web3.fromWei(balance, 'ether'))
+        else:
+            raise ModelRunError(f'Not supported for chain id: {context.chain_id}')
 
 
 class NonFungibleToken(Contract):
@@ -192,3 +265,100 @@ class Tokens(IterableListGenericDTO[Token]):
     tokens: List[Token] = DTOField(
         default=[], description="An iterable list of Token Objects")
     _iterator: str = PrivateAttr('tokens')
+
+
+class FiatCurrency(Account):
+    """
+    This is DTO for Fiat Currency.
+    """
+
+    class FiatCurrencyMeta:
+        symbol: Union[str, None] = None
+        name: Union[str, None] = None
+
+    _meta: FiatCurrencyMeta = PrivateAttr(
+        default_factory=lambda: FiatCurrency.FiatCurrencyMeta())  # pylint: disable=unnecessary-lambda
+
+    def __init__(self, **data):
+        address = data.get('address', None)
+        symbol = data.get("symbol", None)
+
+        if symbol is None and address is None:
+            raise ModelDataError('Missing both symbol and address')
+
+        if symbol is not None:
+            fiat_meta = FIAT_CURRENCY_DATA_BY_SYMBOL.get(symbol, None)
+            if fiat_meta is not None:
+                if address is None or fiat_meta['address'] == address:
+                    super().__init__(address=Address(fiat_meta["address"]))
+                    self._meta.symbol = fiat_meta["symbol"]
+                    self._meta.name = fiat_meta["name"]
+                else:
+                    raise ModelDataError(
+                        f'Mismatch {symbol}/{address} for '
+                        f'{fiat_meta["symbol"]}/{fiat_meta["address"]}')
+            elif address is None:
+                raise ModelDataError(f'{symbol} is not added for fiat currency.')
+
+        if address is not None:
+            fiat_meta = FIAT_CURRENCY_DATA_BY_ADDRESS.get(address, None)
+            if fiat_meta is not None:
+                if symbol is None or fiat_meta['symbol'] == symbol:
+                    super().__init__(address=Address(fiat_meta["address"]))
+                    self._meta.symbol = fiat_meta["symbol"]
+                    self._meta.name = fiat_meta["name"]
+                else:
+                    raise ModelDataError(
+                        f'Mismatch {symbol}/{address} for '
+                        f'{fiat_meta["symbol"]}/{fiat_meta["address"]}')
+            elif symbol is None:
+                raise ModelDataError(f'{address} is not added for fiat currency.')
+            else:
+                raise ModelDataError(f'{symbol}/{address} is not added for fiat currency.')
+
+    @property
+    def symbol(self):
+        return self._meta.symbol
+
+    @property
+    def name(self):
+        return self._meta.name
+
+    @property
+    def fiat(self):
+        return True
+
+
+class Currency(DTO):
+    """
+    This is a converter for any Fungible Token and FiatCurrency.
+    It's used as inputs to price models.
+    """
+
+    address: Union[Address, None] = None
+    symbol: Union[str, None] = None
+    name: Union[str, None] = None
+    fiat: Union[bool, None] = None
+
+    def __new__(cls, **data) -> Union[NativeToken, Token, FiatCurrency]:
+        addr = data.get("address", None)
+        symbol = data.get("symbol", None)
+        fiat = data.get("fiat", None)
+
+        if addr is not None:
+            if (FIAT_CURRENCY_DATA_BY_ADDRESS.get(addr, None) is not None and
+                    (fiat or fiat is None)):
+                return FiatCurrency(**data)
+            if fiat is None or not fiat:
+                return Token(**data)
+
+        if symbol is not None:
+            if (FIAT_CURRENCY_DATA_BY_SYMBOL.get(symbol, None) is not None and
+                    (fiat or fiat is None)):
+                return FiatCurrency(**data)
+            if fiat is None or not fiat:
+                return Token(**data)
+
+        raise ModelDataError(
+            "Could not identify specific currency. Currency "
+            "must be of type Token, NativeToken or FiatCurrency")